Add favourite list and add/remove functionality
A user can now add a joke to a favourite list by clicking on the joke.
This transfers the joke to a second list, which has a remove action on
the click handler which removes the joke from the list.

Adds two new actions, and also refines the type structure of the store
and the selectors. Also adds some very basic CSS.