Add auto-add button which can be activated
Depending on autoAdd state, a timer will fire every x seconds to fetch a
new favourite and add it to the list. This happens until the list is full,
after which the autoAdd state will be set to false again.

Also changed the add/remove functions to a simple toggle and added that to
the favourite list as option as well (remove by clicking again).

The favourite max size is now settable through a simple const;